% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function-Params.R, R/Method-Params.R
\name{CombineParam}
\alias{CombineParam}
\alias{cutoff,CombineParam-method}
\alias{cutoff}
\alias{cutoff<-,CombineParam-method}
\alias{cutoff<-}
\alias{weightRT,CombineParam-method}
\alias{weightRT}
\alias{weightRT<-,CombineParam-method}
\alias{weightRT<-}
\alias{weightCCS,CombineParam-method}
\alias{weightCCS}
\alias{weightCCS<-,CombineParam-method}
\alias{weightCCS<-}
\alias{weightMSMS,CombineParam-method}
\alias{weightMSMS}
\alias{weightMSMS<-,CombineParam-method}
\alias{weightMSMS<-}
\alias{scoreMSMS,CombineParam-method}
\alias{scoreMSMS}
\alias{scoreMSMS<-,CombineParam-method}
\alias{scoreMSMS<-}
\title{Combine score parameter setup}
\usage{
CombineParam(
  cutoff = 0.6,
  weightRT = 0,
  weightCCS = 0,
  weightMSMS = 0,
  scoreMSMS = c("reverse", "forward")
)

\S4method{cutoff}{CombineParam}(object)

\S4method{cutoff}{CombineParam}(object) <- value

\S4method{weightRT}{CombineParam}(object)

\S4method{weightRT}{CombineParam}(object) <- value

\S4method{weightCCS}{CombineParam}(object)

\S4method{weightCCS}{CombineParam}(object) <- value

\S4method{weightMSMS}{CombineParam}(object)

\S4method{weightMSMS}{CombineParam}(object) <- value

\S4method{scoreMSMS}{CombineParam}(object)

\S4method{scoreMSMS}{CombineParam}(object) <- value
}
\arguments{
\item{cutoff}{\code{numeric} threshold for a avaliable match of combined score}

\item{weightRT}{\code{numeric} weight of RT when combining the scors}

\item{weightCCS}{\code{numeric} weight of CCS when combining the scors}

\item{weightMSMS}{\code{numeric} weight of MSMS when combining the scors}

\item{scoreMSMS}{\code{character} MSMS score used for score combination
(only "reverse" or "forward" is supported)}

\item{value}{The value for the slot.}
}
\value{
a \code{CombineParam} object
}
\description{
\code{cutoff},\code{cutoff<-}: getter and setter for the \code{cutoff}
    slot of the object.

\code{weightRT},\code{weightRT<-}: getter and setter for the \code{weightRT}
    slot of the object.

\code{weightCCS},\code{weightCCS<-}: getter and setter for the \code{weightCCS}
    slot of the object.

\code{weightMSMS},\code{weightMSMS<-}: getter and setter for the \code{weightMSMS}
    slot of the object.

\code{scoreMSMS},\code{scoreMSMS<-}: getter and setter for the \code{scoreMSMS}
    slot of the object.
}
