% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function-SearchSpectra.R
\name{.TrapezoidalScore}
\alias{.TrapezoidalScore}
\title{.TrapezoidalScore}
\usage{
.TrapezoidalScore(
  dataExp,
  dataRef,
  tolerance,
  type = c("absolute", "percentage")
)
}
\arguments{
\item{dataExp}{numeric. experimental value}

\item{dataRef}{numeric. referance value}

\item{tolerance}{numeric vector of two. tolerance range of scoring.}
}
\description{
.TrapezoidalScore
}
