% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function-ParseSpectra.R
\name{ParseMSP}
\alias{ParseMSP}
\alias{MSP}
\alias{spectrum}
\alias{file}
\alias{parser}
\alias{ParseMGF}
\alias{ParseCEF}
\alias{CEF}
\alias{ParseTXT}
\alias{TXT}
\alias{ParseCSV}
\alias{CSV}
\alias{ParseMZXML}
\alias{mzXML}
\title{ParseMSP}
\usage{
ParseMSP(file, colSpectra = c(1, 2), nameSpectra = c("mz", "intensity"), ...)

ParseMGF(file, colSpectra = c(1, 2), nameSpectra = c("mz", "intensity"), ...)

ParseCEF(file, colSpectra = c(1, 2), nameSpectra = c("mz", "intensity"), ...)

ParseTXT(
  file,
  colSpectra = c(1, 2, 3),
  nameSpectra = c("mz", "intensity", "intensity.rel"),
  colAnno = c(1, 2),
  nameAnno = c("mzAnno", "formula"),
  ppm = 5,
  ...
)

ParseCSV(
  file,
  colSpectra = c(1, 2),
  nameSpectra = c("mz", "intensity"),
  skip = 0L,
  ...
)

ParseMZXML(file, ...)
}
\arguments{
\item{file}{character: path of the mzXML file}

\item{colSpectra}{vector of two intergers: columns indices recording spectra info}

\item{skip}{integer: the number of lines of the data file to skip before
beginning to read data.}
}
\description{
Read a MSP file and parse it to list of spectra with corresponding
precursor information

Read a MGF file and parse it to list of spectra with corresponding
precursor information

Read a CEF file and parse it to list of spectra with corresponding
precursor information

Read a TXT file of MassBank Euroup and parse it to list of
spectra with corresponding precursor information

Read a CSV file and parse it to list of spectra with corresponding
precursor information

Read a mzXML file and parse it to list of spectra with corresponding
precursor information
}
