% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function-Objects.R
\name{MatchScore}
\alias{MatchScore}
\title{Generate SpectraData object}
\usage{
MatchScore(info, matchedFragments, nlFragments = NULL)
}
\arguments{
\item{info}{\code{data.frame} score information}

\item{matchedFragments}{\code{list} matched fragments info}
}
\description{
Generate SpectraData object
}
