% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Method-Rules.R
\name{ApplyRules,MatchScore-method}
\alias{ApplyRules,MatchScore-method}
\alias{RefineRuleResult,MatchScore-method}
\title{Apply annotation rules}
\usage{
\S4method{ApplyRules}{MatchScore}(matchScore, rules, ...)

\S4method{RefineRuleResult}{MatchScore}(matchScore, ...)
}
\description{
Apply annotation rules

Apply annotation rules
}
