% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Method-Spectra.R
\name{FindSpectra,SpectraData-method}
\alias{FindSpectra,SpectraData-method}
\alias{FilterFragments,SpectraData-method}
\alias{FilterNULL,SpectraData-method}
\title{Find spectra
Find spectra with user defined terms}
\usage{
\S4method{FindSpectra}{SpectraData}(spectra, findby = NULL, values = NULL, colSpecName = NULL)

\S4method{FilterFragments}{SpectraData}(
  spectra,
  mzrange = NULL,
  thrIntensityAbs = NULL,
  thrIntensityRel = NULL
)

\S4method{FilterNULL}{SpectraData}(spectra, keepInfo = TRUE)
}
\arguments{
\item{spectra}{\code{SpectraData} Spectra object for the spectra to be found}

\item{findby}{\code{character} Terms for spectra finding}

\item{mzrange}{\code{numeric(2)} Fragment mass range}

\item{thrIntensityAbs}{\code{numeric} Absolute fragment intensity threshold to be kept}

\item{thrIntensityRel}{\code{numeric} Relative fragment intensity threshold to be kept}

\item{value}{\code{character} Values of terms for spectra finding}
}
\value{
a \code{SpectraData} object

a \code{SpectraData} object

a \code{SpectraData} object
}
\description{
Find spectra
Find spectra with user defined terms

Filter fragments
Filtering fragments in SpectraData with mass range and fragment intensity

Filter NULL spectra
Filter the NULL spectra from SpectraData
}
