% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Method-ParseSpectra.R
\name{ParseSpectra,ParseSpectraParam-method}
\alias{ParseSpectra,ParseSpectraParam-method}
\alias{Parse}
\alias{spectra}
\title{ParseSpectra}
\usage{
\S4method{ParseSpectra}{ParseSpectraParam}(parseParam, files, ...)
}
\arguments{
\item{parseParam}{ParamParseSpectra object, recording parameters for parsing
spectra}

\item{files}{character: path of the spectrum files}
}
\description{
Get spectra data by reading spectrum files
}
