% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function-Params.R, R/Method-Params.R
\name{SearchParam}
\alias{SearchParam}
\alias{ppm,SearchParam-method}
\alias{ppm}
\alias{ppm<-,SearchParam-method}
\alias{ppm<-}
\alias{scoreRT,SearchParam-method}
\alias{scoreRT}
\alias{scoreRT<-,SearchParam-method}
\alias{scoreRT<-}
\alias{toleranceRT,SearchParam-method}
\alias{toleranceRT}
\alias{toleranceRT<-,SearchParam-method}
\alias{toleranceRT<-}
\alias{scoreCCS,SearchParam-method}
\alias{scoreCCS}
\alias{scoreCCS<-,SearchParam-method}
\alias{scoreCCS<-}
\alias{toleranceCCS,SearchParam-method}
\alias{toleranceCCS}
\alias{toleranceCCS<-,SearchParam-method}
\alias{toleranceCCS<-}
\alias{typeCCS,SearchParam-method}
\alias{typeCCS}
\alias{typeCCS<-,SearchParam-method}
\alias{typeCCS<-}
\alias{adductIncluded,SearchParam-method}
\alias{adductIncluded}
\alias{adductIncluded<-,SearchParam-method}
\alias{adductIncluded<-}
\alias{adductExcluded,SearchParam-method}
\alias{adductExcluded}
\alias{adductExcluded<-,SearchParam-method}
\alias{adductExcluded<-}
\alias{adductFile,SearchParam-method}
\alias{adductFile}
\alias{adductFile<-,SearchParam-method}
\alias{adductFile<-}
\alias{classIncluded,SearchParam-method}
\alias{classIncluded}
\alias{classIncluded<-,SearchParam-method}
\alias{classIncluded<-}
\alias{classExcluded,SearchParam-method}
\alias{classExcluded}
\alias{classExcluded<-,SearchParam-method}
\alias{classExcluded<-}
\alias{resDefineAt,SearchParam-method}
\alias{resDefineAt}
\alias{resDefineAt<-,SearchParam-method}
\alias{resDefineAt<-}
\title{Searching spectra Parameters}
\usage{
SearchParam(
  ppm = 25,
  scoreRT = TRUE,
  toleranceRT = c(32, 64),
  scoreCCS = TRUE,
  toleranceCCS = c(1.34, 2.68),
  typeCCS = c("percentage", "absolute"),
  adductIncluded = NULL,
  adductExcluded = NULL,
  adductFile = NULL,
  classIncluded = NULL,
  classExcluded = NULL,
  useMS1ResDefine = TRUE,
  updateRefMZ = TRUE,
  resDefineAt = 400
)

\S4method{ppm}{SearchParam}(object)

\S4method{ppm}{SearchParam}(object) <- value

\S4method{scoreRT}{SearchParam}(object)

\S4method{scoreRT}{SearchParam}(object) <- value

\S4method{toleranceRT}{SearchParam}(object)

\S4method{toleranceRT}{SearchParam}(object) <- value

\S4method{scoreCCS}{SearchParam}(object)

\S4method{scoreCCS}{SearchParam}(object) <- value

\S4method{toleranceCCS}{SearchParam}(object)

\S4method{toleranceCCS}{SearchParam}(object) <- value

\S4method{typeCCS}{SearchParam}(object)

\S4method{typeCCS}{SearchParam}(object) <- value

\S4method{adductIncluded}{SearchParam}(object)

\S4method{adductIncluded}{SearchParam}(object) <- value

\S4method{adductExcluded}{SearchParam}(object)

\S4method{adductExcluded}{SearchParam}(object) <- value

\S4method{adductFile}{SearchParam}(object)

\S4method{adductFile}{SearchParam}(object) <- value

\S4method{classIncluded}{SearchParam}(object)

\S4method{classIncluded}{SearchParam}(object) <- value

\S4method{classExcluded}{SearchParam}(object)

\S4method{classExcluded}{SearchParam}(object) <- value

\S4method{resDefineAt}{SearchParam}(object)

\S4method{resDefineAt}{SearchParam}(object) <- value
}
\arguments{
\item{ppm}{\code{numeric} ppm tolerance for searching precursor mz}

\item{scoreRT}{\code{logical} if comparing RT}

\item{toleranceRT}{\code{numeric (2)} c(Tmin, Tmax), RT range for trapezoidal score.
\itemize{
   \item[*] Tmin: Topline for trapezoidal score
   \item[*] Tmax: Baseline for trapezoidal score
 }}

\item{scoreCCS}{\code{logical} if comparing CCS}

\item{toleranceCCS}{\code{numeric (2)} c(Cmin, Cmax), CCS range for trapezoidal
 score.
\itemize{
   \item[*] Cmin: Topline for trapezoidal score
   \item[*] Cmax: Baseline for trapezoidal score
 }}

\item{typeCCS}{\code{character} CCS tolerance type, either "percentage" or
"absolute"}

\item{adductIncluded}{\code{character} adduct types to be included in MS1 match}

\item{adductFile}{\code{character} file path for user provided adduct table}

\item{classIncluded}{\code{character} compound classes to be included in MS1 match}

\item{useMS1ResDefine}{\code{logical} if use resDefineAt when calculating precursor m/z tolerance}

\item{updateRefMZ}{\code{logical} if update reference precursor m/z with it's corresponding
adduct type}

\item{resDefineAt}{\code{numeric} m/z (Da) value for resolution definition}

\item{value}{The value for the slot.}

\item{adductExclude}{\code{character} adduct types to be excluded in MS1 match}

\item{classExclude}{\code{character} compound classes to be excluded in MS1 match}
}
\description{
Parameters for searching experimentall related SpectraData from refrence
SpectraData

\code{ppm},\code{ppm<-}: getter and setter for the \code{ppm}
    slot of the object.

\code{scoreRT},\code{scoreRT<-}: getter and setter for the \code{scoreRT}
    slot of the object.

\code{toleranceRT},\code{toleranceRT<-}: getter and setter for the \code{toleranceRT}
    slot of the object.

\code{scoreCCS},\code{scoreCCS<-}: getter and setter for the \code{scoreCCS}
    slot of the object.

\code{toleranceCCS},\code{toleranceCCS<-}: getter and setter for the \code{toleranceCCS}
    slot of the object.

\code{typeCCS},\code{typeCCS<-}: getter and setter for the \code{typeCCS}
    slot of the object.

\code{adductIncluded},\code{adductIncluded<-}: getter and setter for the \code{adductIncluded}
    slot of the object.

\code{adductExcluded},\code{adductExcluded<-}: getter and setter for the \code{adductExcluded}
    slot of the object.

\code{adductFile},\code{adductFile<-}: getter and setter for the \code{adductFile}
    slot of the object.

\code{classIncluded},\code{classIncluded<-}: getter and setter for the \code{classIncluded}
    slot of the object.

\code{classExcluded},\code{classExcluded<-}: getter and setter for the \code{classExcluded}
    slot of the object.

\code{resDefineAt},\code{resDefineAt<-}: getter and setter for the \code{resDefineAt}
    slot of the object.
}
