% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function-Objects.R
\name{SpectraData}
\alias{SpectraData}
\title{Generate SpectraData object}
\usage{
SpectraData(info, spectra, colSpecName = NULL)
}
\arguments{
\item{info}{\code{data.frame} metabolite information}

\item{spectra}{\code{list} spectra of metabolite.}
}
\description{
Generate SpectraData object
}
