% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{SpectraTools}
\alias{SpectraTools}
\title{SpectraTools}
\description{
Dealing with spectra
}
\details{
This package processes MS spectra data including parsing spectra recorded in
different types of files, searching and matching MSMS spectra.
}
\author{
Yandong Yin (\email{yinyandong@sioc.ac.cn})
}
